requirements.txt
jupyterlab==3.6.3
ipywidgets==8.0.6
matplotlib==3.7.1
seaborn==0.12.2
numpy==1.24.3
pandas==2.0.1
pyarrow==14.0.1
scikit-learn==1.2.2
statsmodels==0.13.5
pytest==7.3.1
missingno==0.5.2
Jinja2==3.1.2
imbalanced-learn==0.11.0
pyspark
apache-airflow
memory_profiler==0.61.0
boto3==1.34.2
python-dotenv
mlflow
